Add BlockGarage
This block is meant for internal testing and as an example of a working implementation. This is not meant for use in a real-world application (hence the iron block texture). You SHOULD make or use a real implementation of this feature. A default one might come in the future, but is not a priority as of now.